requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.2
django-colorfield==0.11.0
django-cors-headers==4.5.0
django-grappelli==4.0.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
oscrypto==1.3.0
pdfkit==1.0.0
pillow==11.0.0
pycparser==2.22
pyHanko==0.25.2
pyhanko-certvalidator==0.26.4
PyJWT==2.9.0
pypdf==5.1.0
python-bidi==0.6.3
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
xhtml2pdf==0.2.16
gunicorn>=20.1.0
